refactor(xo-server): extract vmware import method to its own mixin
